// app/generated/api/models/branchRequest.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Pricing Backend API
 * REST API for managing branches, regions, products, and pricing plans.
 * OpenAPI spec version: 1.0.0
 */
import type { BranchCode } from './branchCode';
import type { State } from './state';
import type { ZipCode } from './zipCode';
import type { UpdatedBy } from './updatedBy';

export interface BranchRequest {
  branchCode: BranchCode;
  /**
   * @minLength 1
   * @maxLength 100
   */
  branchName: string;
  state: State;
  zipCode: ZipCode;
  updatedBy: UpdatedBy;
}